+++ src/main/java/exception/AccountNotFound.java
@@ -1,7 +1,7 @@
 package exception;
 
 public class AccountNotFound extends RuntimeException {
-    public AccountNotFound(int id) {
+    public AccountNotFound(Long id) {
         super("Account not found with id: " + id);
     }
 }
